Move stuff to their own files
Bonus: increase quality of recordings SO MUCH
