+++ app/src/main/java/com/hedvig/app/feature/embark/passages/audiorecorder/AudioRecorderParameters.kt
@@ -1,0 +1,9 @@
+package com.hedvig.app.feature.embark.passages.audiorecorder
+
+import android.os.Parcelable
+import kotlinx.parcelize.Parcelize
+
+@Parcelize
+data class AudioRecorderParameters(
+    val messages: List<String>,
+) : Parcelable
